Move SmallCharSet to markup5ever
